refactor(SmartEditor): extract generic Slate+LSP primitive from DpqlEditor
Split the monolithic DpqlEditor into a generic language-agnostic
primitive (`SmartEditor`) and a thin DPQL-flavored wrapper. The
underlying machinery — Slate editor wrapping `ReqoreRichTextEditor`,
LSP session lifecycle, trigger-character detection, dropdown rendering,
keyboard navigation, diagnostic flow — is now language-neutral. Future
work to build `QonsoleSmartInput` (qorus-ide issue #196 task #3) can
reuse the same primitive by pointing it at the Qonsole LSP, which the
live spike against `wss://hq.qoretechnologies.com:8092/lsp` (captured
in QONSOLE_LSP_REFERENCE.md and QONSOLE_LSP_RESPONSES.txt at the
qorus-frontend root) confirmed is already shipped.

New `src/components/smartEditor/`:

- `SmartEditor.tsx` — generic Slate editor primitive. Takes a
  caller-owned `session: IUseLspSessionResult`, plus language-flavor
  props (`decorate`, `customRenderLeaf`, `tagRenderer`,
  `triggerCharacters`, `converter`, `completionInserter`, `topActions`).
  Forwards the underlying Slate editor ref.
- `useLspSession.ts` — generic LSP session lifecycle. Connect → didOpen
  with optional metadata → forward didChange → expose `client`, `uri`,
  `isReady`, `diagnostics`, `didChange`, `getCompletions`, `format`.
  Lifting the session out of SmartEditor lets wrappers subscribe to its
  state in effects (e.g. DpqlEditor calls `dpql/setContext` whenever
  `session.client` + `provider` are both available).
- `useLspAutocomplete.ts` — generic dropdown UI hook. Takes the
  converter and inserter as props; was hardcoded to DPQL behavior in
  the pre-refactor `useDpqlAutocomplete`.
- `helpers.ts` — generic defaults for `toSlateNodes` /
  `fromSlateNodes` / `selectionToOffset`, plus the language-agnostic
  `offsetToLspPosition` and `mapCompletionKindToIcon`.
- `types.ts` — `ISmartEditorProps`, `ISlateConverter`,
  `TCompletionInserter`, generic Slate node types.

DPQL becomes a wrapper in `src/components/dpqlEditor/`:

- `DpqlEditor.tsx` — thin: lays out the templates dropdown above the
  editor, wires DPQL-flavored knobs (`dpqlSlateConverter`,
  `dpqlCompletionInserter`, syntax highlighting via the existing
  `useDpqlSyntaxHighlighting`, DPQL tag renderer + colors), and exposes
  the typed `format` / `validate` / `parse` / `serialize` ref API
  backed by `dpql/*` custom-method calls.
- `useDpqlSession.ts` — DPQL layer on top of `useLspSession`. Reacts
  to `provider` / `recordType` / `options` / `actionCode` changes by
  dispatching `dpql/setContext` and tracking the returned `fieldMeta`.
  Wraps `dpql/parse` / `dpql/serialize` / `dpql/validate` as typed
  Promise-returning functions.
- `dpqlHelpers.ts` — DPQL-aware `plainTextToSlate` /
  `slateToPlainText` / `slateSelectionToOffset` (handles
  `metadata.quoted` tag spacing) / `isTagCompletion` / `getTagLabel` +
  the `dpqlSlateConverter` packaged for SmartEditor.
- `dpqlInserter.ts` — completion inserter that drops a Slate tag
  element for `@field` / `$template` completions and falls back to
  plain text for keywords / operators.
- `dpqlTags.ts` — color scheme per template prefix + `makeDpqlTagRenderer`
  closure factory (so `@field` tooltips reflect the live `fieldMeta`
  from `useDpqlSession`).

Tests:

- `__tests__/smartEditor/helpers.test.ts` — 17 new tests for
  generic helpers (`offsetToLspPosition`, `defaultToSlateNodes`,
  `defaultFromSlateNodes`, `defaultSelectionToOffset`,
  `mapCompletionKindToIcon`).
- `__tests__/dpqlEditor/helpers.test.ts` — DPQL-specific tests stay,
  `offsetToLspPosition` tests moved to the smartEditor suite.

The public DpqlEditor surface (`IDpqlEditorProps`, `IDpqlEditorRef`,
component name, ref methods) is unchanged — the existing Storybook
stories and the `LspCompletionRoundtrip` play test continue to pass.

Pre-push checks all green: TS clean, lint clean, 246 / 246 tests
across 7 suites (up from 229/6), `yarn build-storybook` succeeds.
