+++ src/components/dpqlEditor/index.ts
@@ -10,17 +10,25 @@ export type {
   ISlateText,
   TSlateNode,
 } from './types';
+export { useDpqlSession } from './useDpqlSession';
 export type {
   IDpqlFieldMeta,
-  IUseDpqlLspOptions,
-  IUseDpqlLspResult,
-} from './useDpqlLsp';
-export { useDpqlLsp } from './useDpqlLsp';
-export type {
-  ICompletionDropdownItem,
-  ICompletionGroup,
-  IUseDpqlAutocompleteOptions,
-  IUseDpqlAutocompleteResult,
-} from './useDpqlAutocomplete';
-export { useDpqlAutocomplete } from './useDpqlAutocomplete';
+  IUseDpqlSessionOptions,
+  IUseDpqlSessionResult,
+} from './useDpqlSession';
 export { useDpqlSyntaxHighlighting } from './useDpqlSyntaxHighlighting';
+export {
+  dpqlSlateConverter,
+  getTagLabel,
+  isTagCompletion,
+  plainTextToSlate,
+  slateSelectionToOffset,
+  slateToPlainText,
+} from './dpqlHelpers';
+export {
+  DEFAULT_TEMPLATE_COLOR,
+  FIELD_REF_COLOR,
+  TEMPLATE_COLORS,
+  makeDpqlTagRenderer,
+} from './dpqlTags';
+export { dpqlCompletionInserter } from './dpqlInserter';
